refactor: 경험 보드 상세 조회 로직 분리
